Move AnimData to generate_config.py
